Day 4/Web Research Agent using ReAct pattern/requirements.txt
streamlit
google-generativeai
tavily-python
reportlab
python-dotenv